refactor: promote validation code to utils
